Modify object key generation fix off-by-1 issue
This commits changes the way we an object key is generated. Instead
of using the all 16 bytes as key, we calculate hash value for the
entire file path and assign it to the 8 high bytes.
In addition, this commits fixes an issue with reading an object that
have the extact same size as the read buffer.

Trello-Card: https://trello.com/c/dZvxCgoj

Co-authored-by: Amit Shoham <amitshoham8950@gmail.com>
Change-Id: If18e213f3783140367f2d8fdadbc8eb6254c9b54