Modify object key generation and fix off-by-1 issue
This change modifies the way object keys are generated. Previously,
the file path was split into two halves: the hash of the first half
was stored in the high bytes, and the hash of the second half was
stored in the low bytes. Now, the entire path is used to calculate
the hash for both the high and low bytes, resulting in the same
hash for both. This change ensures the use of 8-byte keys instead
of 16-byte keys, as required by the OSD.
In addition, this change addresses an issue with reading objects
that have the exact same size as the read buffer.

Trello-Card: https://trello.com/c/dZvxCgoj

Co-authored-by: Amit Shoham <amitshoham8950@gmail.com>
Change-Id: If18e213f3783140367f2d8fdadbc8eb6254c9b54